+++ src/Service/Admin/LicenseManagement/LicenseManagementManager.php
@@ -4,14 +4,25 @@ namespace WP_Statistics\Service\Admin\LicenseManagement;
 
 use Exception;
 use WP_Statistics;
-use WP_STATISTICS\User;
-use WP_STATISTICS\Option;
 use WP_Statistics\Exception\LicenseException;
-use WP_Statistics\Service\Admin\NoticeHandler\Notice;
+use WP_STATISTICS\Option;
 use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginActions;
 use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginHandler;
 use WP_Statistics\Service\Admin\LicenseManagement\Plugin\PluginUpdater;
+use WP_Statistics\Service\Admin\NoticeHandler\Notice;
+use WP_STATISTICS\User;
 
+/**
+ * LicenseManagementManager Class
+ *
+ * Manages the activation and validation of licenses for WP Statistics add-ons.
+ * Provides methods to handle license keys and ensure proper functionality of premium features.
+ *
+ * @package   LicenseManagement
+ * @version   1.0.0
+ * @since     14.0
+ * @author    Foad
+ */
 class LicenseManagementManager
 {
     /** @var ApiCommunicator */