+++ src/Service/Charts/DataProvider/ModelChartDataProvider.php
@@ -2,6 +2,7 @@
 
 namespace WP_Statistics\Service\Charts\DataProvider;
 
+use WP_STATISTICS\Admin_Template;
 use WP_Statistics\Models\VisitorsModel;
 use WP_Statistics\Service\Charts\AbstractChartDataProvider;
 use WP_Statistics\Service\Charts\Traits\BarChartResponseTrait;
@@ -17,8 +18,22 @@ class ModelChartDataProvider extends AbstractChartDataProvider
         parent::__construct($args);
 
         $this->args = array_merge($this->args, [
-            'fields' => ['visitor.model']
+            'fields'   => ['model' => 'visitor.model', 'visitors' => 'COUNT(visitor.ID) as visitors'],
+            'group_by' => 'visitor.model',
+            'order_by' => 'visitors',
+            'decorate' => false,
+            'page'     => false,
+            'per_page' => false
         ]);
+
+        // If filter is applied, get distinct visitors to avoid data duplication
+        if ($this->isFilterApplied()) {
+            $this->args['fields']['visitors'] = 'COUNT(DISTINCT visitor.ID) as visitors';
+        }
+
+        if (empty($this->args['limit'])) {
+            $this->args['limit'] = 5;
+        }
 
         $this->visitorsModel = new VisitorsModel();
     }
@@ -28,7 +43,12 @@ class ModelChartDataProvider extends AbstractChartDataProvider
     {
         $this->initChartData();
 
-        $data = $this->visitorsModel->getVisitorsData($this->args);
+        if (!empty($this->args['referred_visitors'])) {
+            $data = $this->visitorsModel->getReferredVisitors($this->args);
+        } else {
+            $data = $this->visitorsModel->getVisitorsData($this->args);
+        }
+
         $data = $this->parseData($data);
 
         $this->setChartLabels($data['labels']);
@@ -40,47 +60,24 @@ class ModelChartDataProvider extends AbstractChartDataProvider
     protected function parseData($data)
     {
         $parsedData = [];
-        $unknownData = 0;
 
         if (!empty($data)) {
             foreach ($data as $item) {
-                $model = $item->getDevice()->getModel();
+                $model = Admin_Template::unknownToNotSet($item->model);
 
-                if (!empty($model)) {
-                    $models = array_column($parsedData, 'label');
-
-                    if (!in_array($model, $models)) {
-                        $parsedData[] = [
-                            'label'    => $model,
-                            'visitors' => 1
-                        ];
-                    } else {
-                        $index = array_search($model, $models);
-                        $parsedData[$index]['visitors']++;
-                    }
-                }
-
-                if ($model === 'Unknown' || empty($model)) {
-                    ++$unknownData;
-                }
-            }
-
-            if ($unknownData > 0) {
                 $parsedData[] = [
-                    'label'    => esc_html__('Unknown', 'wp-statistics'),
-                    'visitors' => $unknownData
+                    'label'    => $model,
+                    'visitors' => $item->visitors
                 ];
             }
 
-            // Sort data by visitors
-            usort($parsedData, function ($a, $b) {
-                return $b['visitors'] - $a['visitors'];
-            });
+            // Limit the number of items. If limit is 5, limit items to 4 + other
+            $limit = $this->args['limit'] - 1;
 
-            if (count($parsedData) > 4) {
+            if (count($parsedData) > $limit) {
                 // Get top 4 results, and others
-                $topData    = array_slice($parsedData, 0, 4);
-                $otherData  = array_slice($parsedData, 4);
+                $topData    = array_slice($parsedData, 0, $limit);
+                $otherData  = array_slice($parsedData, $limit);
 
                 // Show the rest of the results as others, and sum up the visitors
                 $otherItem    = [