+++ src/Service/Database/DatabaseFactory.php
@@ -19,6 +19,11 @@ use WP_Statistics\Service\Database\Operations\Update;
  *
  * This class provides methods to create specific operations (e.g., create, update, drop)
  * and manage different migration types (e.g., schema, data).
+ *
+ * @package   Database
+ * @version   1.4.0
+ * @since     14.12.3
+ * @author    Hooman
  */
 class DatabaseFactory
 {
@@ -48,6 +53,13 @@ class DatabaseFactory
     ];
 
     /**
+     * Cache of instantiated table operation.
+     *
+     * @var array<string, AbstractTableOperation>
+     */
+    private static $operationInstance = [];
+
+    /**
      * Create an instance of a specific table operation.
      *
      * @param string $operation The name of the operation (e.g., 'create', 'drop').
@@ -57,6 +69,10 @@ class DatabaseFactory
     public static function table($operation)
     {
         $operation = strtolower($operation);
+
+        if (!empty(self::$operationInstance[$operation])) {
+            return self::$operationInstance[$operation];
+        }
 
         if (!isset(self::$operations[$operation])) {
             throw new \InvalidArgumentException("Invalid operation: {$operation}");
@@ -68,7 +84,9 @@ class DatabaseFactory
             throw new \InvalidArgumentException("Class not exist: {$providerClass}");
         }
 
-        return new $providerClass();
+        self::$operationInstance[$operation] = new $providerClass();
+
+        return self::$operationInstance[$operation];
     }
 
     /**