+++ views/components/objects/google-data-policy-alert.php
@@ -1,6 +1,6 @@
 <div class="wps-addon-settings-marketing--alert">
     <div>
         <h3><?php esc_html_e('Privacy & Data', 'wp-statistics'); ?></h3>
-        <p><?php esc_html_e('We do not store any of the data from your Google account on our servers, everything is processed & stored on your server. We take your privacy extremely seriously and ensure it is never misused.', 'wp-statistics'); ?> <a href="<?php echo esc_url(WP_STATISTICS_SITE_URL . '/resources/google-search-console-integration-privacy-data-handling/?utm_source=wp-statistics&utm_medium=link&utm_campaign=settings') ?>" target="_blank"><?php esc_html_e('Learn more', 'wp-statistics'); ?></a>.</p>
+        <p><?php echo wp_kses_post($content); ?></p>
     </div>
 </div>