+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+= 14.16 - 2025-x-x =
+- **Enhancement:** Removed Matomo Referrer Spam Blacklist due to misalignment, performance concerns, and functionality issues.
+
+= 14.15.x - 2025-xx-xx =
+- **Fix:** Prevented wp_statistics_tracker_js_errors option bloat by limiting the maximum number of saved errors and truncating error messages.
+- **Fix:** Fixed `Top Pages` widget sort issue on `Page Insights` report.
+- **Fix:** Region widget issue when country is not detected on `Geographic` page.
+- **Fix:** Prevented false `Queue Migration` notice after updates.
+- **Enhancement:** Refactored the update process and database schema updates to run on the frontend.
+- **Enhancement:** Optimized the database manager to prevent duplicate queries.
+
 = 14.15.3 - 2025-08-18 =
 - **Fix:** Fixed a warning by validating geographic location codes are strings or integers before use.
 - **Fix:** Correct redirect URL after clicking 'Start Migration'.